requirement.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
lightgbm
catboost
joblib
xgboost
streamlit
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage
environs